profile_server/requirement.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.6.1
pytz==2018.3
wrapt==1.10.11
